feat: more refactoring
