+++ packages/crane/src/lib/utils/index.js
@@ -3,6 +3,7 @@ export { default as createSvgIcon } from './createSvgIcon';
 export { default as insertLink } from './insertLink';
 export { default as getIntentColor } from './getIntentColor';
 export { default as getContrastColor } from './getContrastColor';
+export { default as parseJSON } from './parseJSON';
 export { default as randomString } from './randomString';
 export { default as separateUnit } from './separateUnit';
 export { default as SvgIcon } from './SvgIcon';
