+++ packages/gatsby-starter-hummingbird/package.json
@@ -3,7 +3,7 @@
   "name": "@wingscms/gatsby-starter-hummingbird",
   "version": "0.1.21",
   "scripts": {
-    "start": "rm -rf .cache && gatsby develop -p 4000",
+    "start": "gatsby clean && gatsby develop -p 4000",
     "build": "gatsby build"
   },
   "dependencies": {
