+++ packages/hummingbird/src/cards/CTACard.js
@@ -1,1 +1,1 @@
-export { CTACard as default } from '@wingscms/hummingbird';
+export { CTACard as default } from '@wingscms/react';
