+++ packages/hummingbird/src/cards/CampaignCard.js
@@ -1,1 +1,5 @@
-export { CampaignCard as default } from '@wingscms/hummingbird';
+import React from 'react';
+import CampaignCard from '@wingscms/react';
+import wings from '../data/wings';
+
+export default props => <CampaignCard {...props} wings={wings} />;
