+++ packages/hummingbird/src/cards/QACard.js
@@ -1,1 +1,1 @@
-export { QACard as default } from '@wingscms/hummingbird';
+export { QACard as default } from '@wingscms/react';
