+++ packages/hummingbird/src/cards/TestimonialCard.js
@@ -1,1 +1,1 @@
-export { TestimonialCard as default } from '@wingscms/hummingbird';
+export { TestimonialCard as default } from '@wingscms/react';
