+++ packages/hummingbird/src/templates/components/Campaign/Campaign.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import styled, { withTheme } from 'styled-components';
 import { Campaign as CampaignMain } from '@wingscms/react';
+import wings from '../../../data/wings';
 
 import Container from '../../../components/Container';
 import LayoutDefault from '../../../components/LayoutDefault';
@@ -103,7 +104,7 @@ class CampaignTemplate extends Component {
     children: [
       <CampaignTemplate.Navigation />,
       <CampaignTemplate.Header />,
-      <CampaignTemplate.Main />,
+      <CampaignTemplate.Main wings={wings} />,
     ],
     layout: LayoutDefault,
   };
